Fixes table helper #table_actions method
A typus resource actions may be configured with a proc which
determines if an action should be rendered for a specific instance.

In the case of a false result #table_actions would return a nil value
which would result in an exception when rendering the view.

This fix this issue #compact is called to remove nil values from the
result array.